pyproject.toml
[project]
name = "projektn5"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "mysql-connector-python>=9.2.0",
    "pytest>=8.3.5",
]
